+++ plugins/06_WarlordsOfDraenor/CHANGELOG.md
@@ -12,7 +12,10 @@
   * Rares for [Gorgrond Monster Hunter](https://www.wowhead.com/achievement=9400/gorgrond-monster-hunter) achievement
   * Rares for [Fight the Power](https://www.wowhead.com/achievement=9655/fight-the-power) achievement
   * Rares for [Ancient No More](https://www.wowhead.com/achievement=9678/ancient-no-more) achievement
+  * Pet Battle for [Taming Draenor](https://www.wowhead.com/achievement=9724/taming-draenor) achievement
   * Treasures for [Grand Treasure Hunter](https://www.wowhead.com/achievement=9728/grand-treasure-hunter) achievement
+  * Pepe for [I Found Pepe!](https://www.wowhead.com/achievement=10053/i-found-pepe) achievement
+  * [In Plain Sight](https://www.wowhead.com/achievement=9656/in-plain-sight) achievement
   * Toys / Pets / Transmogs